+++ bot/classes/maps.py
@@ -71,6 +71,18 @@ class MapSelection():
         self.__status = SelStatus.IS_EMPTY
         _mapSelectionsDict[self.__id] = self
 
+    @classmethod
+    def newFromId(cls, matchId, mapId):
+        obj = cls(matchId)
+        obj.__id = matchId
+        i = 0
+        while _allMapsList[i].id != mapId:
+            i+=1
+        obj.__selection = [_allMapsList[i]]
+        obj.__selected = _allMapsList[i]
+        obj.__status = SelStatus.IS_CONFIRMED
+        return obj
+
     def __doSelection(self, args):
         if self.__status is SelStatus.IS_SELECTION and len(args) == 1 and args[0].isnumeric():
             index = int(args[0])