+++ bot/classes/players.py
@@ -49,23 +49,23 @@ class Player():
     """ Basic player class, every registered user matches a Player object contained in the dictionary
     """
 
-    def __init__(self, name, id):
+    def __init__(self, id, name):
         self.__name = name
         self.__id = id
         self.__rank = 0
-        self._igNames = ["N/A", "N/A", "N/A"]
-        self._igIds = [0, 0, 0]
+        self.__igNames = ["N/A", "N/A", "N/A"]
+        self.__igIds = [0, 0, 0]
         self.__notify = False
         self.__timeout = {"time" : 0, "reason" :""}
         self.__status = PlayerStatus.IS_NOT_REGISTERED
-        self._hasOwnAccount = False
-        self._active = None
-        self._match = None
+        self.__hasOwnAccount = False
+        self.__active = None
+        self.__match = None
         _allPlayers[id] = self  # Add to dictionary on creation
 
     @classmethod
     def newFromData(cls, data):  # make a new Player object from database data
-        obj = cls(data["name"], data["_id"])
+        obj = cls(data["_id"], data["name"])
         obj.__rank = data["rank"]
         if obj.__rank == 0:
             obj.__status = PlayerStatus.IS_NOT_REGISTERED
@@ -73,25 +73,25 @@ class Player():
             obj.__status = PlayerStatus.IS_REGISTERED
         obj.__notify = data["notify"]
         obj.__timeout = data["timeout"]
-        obj._igNames = data["igNames"]
-        obj._igIds = data["igIds"]
-        obj._hasOwnAccount = data["hasOwnAccount"]
-        for i in range(len(obj._igIds)):
-            _namesChecking[i][obj._igIds[i]] = obj
+        obj.__igNames = data["igNames"]
+        obj.__igIds = data["igIds"]
+        obj.__hasOwnAccount = data["hasOwnAccount"]
+        for i in range(len(obj.__igIds)):
+            _namesChecking[i][obj.__igIds[i]] = obj
         return obj
 
     async def dbUpdate(self, arg):
         if arg == "notify":
             await updatePlayer(self, {"notify": self.__notify})
         elif arg == "register":
-            await updatePlayer(self, {"igNames": self._igNames, "igIds": self._igIds,
-                                      "rank": self.__rank, "hasOwnAccount": self._hasOwnAccount})
+            await updatePlayer(self, {"igNames": self.__igNames, "igIds": self.__igIds,
+                                      "rank": self.__rank, "hasOwnAccount": self.__hasOwnAccount})
         elif arg == "timeout":
             await updatePlayer(self, {"timeout": self.__timeout})
 
     @property
     def active(self):  # "Active player" object, when player is in a match, contains more info
-        return self._active
+        return self.__active
 
     @property
     def name(self):
@@ -105,6 +105,11 @@ class Player():
     def timeout(self, time, reason=""):
         self.__timeout["time"] = time
         self.__timeout["reason"] = reason
+
+    # TODO
+    # DEV
+    def cheatName(self, name):
+        self.__igNames = [name, name, name]
 
     @property
     def isTimeout(self):
@@ -146,19 +151,20 @@ class Player():
         self.__status = PlayerStatus.IS_PLAYING
 
     def onPlayerClean(self):
-        self._match = None
-        self._active = None
-        if not self._hasOwnAccount:
-            self._igNames = ["N/A", "N/A", "N/A"]
-            self._igIds = [0, 0, 0]
+        self.__match = None
+        self.__active = None
+        if not self.__hasOwnAccount:
+            self.__igNames = ["N/A", "N/A", "N/A"]
+            self.__igIds = [0, 0, 0]
         self.__status = PlayerStatus.IS_REGISTERED
         self.updateRole()
 
-    def onPicked(self):
+    def onPicked(self, active):
+        self.__active = active
         self.__status = PlayerStatus.IS_PICKED
 
     def onMatchSelected(self, m):
-        self._match = m
+        self.__match = m
         self.__status = PlayerStatus.IS_MATCHED
         self.inactiveTask.cancel()
 
@@ -171,7 +177,7 @@ class Player():
             self.inactiveTask.cancel()
 
     def onResign(self):
-        self._active = None
+        self.__active = None
         self.__status = PlayerStatus.IS_MATCHED
 
     @tasks.loop(minutes=cfg.AFK_TIME, delay=1, count=2)
@@ -196,11 +202,11 @@ class Player():
 
     @property
     def igNames(self):
-        return self._igNames
+        return self.__igNames
 
     @property
     def igIds(self):
-        return self._igIds
+        return self.__igIds
 
     @property
     def status(self):
@@ -208,11 +214,15 @@ class Player():
 
     @property
     def match(self):  # when in match
-        return self._match
+        return self.__match
 
     @property
     def hasOwnAccount(self):
-        return self._hasOwnAccount
+        return self.__hasOwnAccount
+
+    def copyIgInfo(self, player):
+        self.__igNames = player.igNames.copy()
+        self.__igIds = player.igIds.copy()
 
     def getData(self):  # get data for database push
         data = {"_id": self.__id,
@@ -220,9 +230,9 @@ class Player():
                 "rank": self.__rank,
                 "notify": self.__notify,
                 "timeout": self.__timeout,
-                "igNames": self._igNames,
-                "igIds": self._igIds,
-                "hasOwnAccount": self._hasOwnAccount
+                "igNames": self.__igNames,
+                "igIds": self.__igIds,
+                "hasOwnAccount": self.__hasOwnAccount
                 }
         return data
 
@@ -232,21 +242,21 @@ class Player():
         """
         updated = False
         if charList is None:
-            if(self.__status is PlayerStatus.IS_NOT_REGISTERED or self._hasOwnAccount):
+            if(self.__status is PlayerStatus.IS_NOT_REGISTERED or self.__hasOwnAccount):
                 updated = True
-            self._igIds = [0, 0, 0]
-            self._igNames = ["N/A", "N/A", "N/A"]
+            self.__igIds = [0, 0, 0]
+            self.__igNames = ["N/A", "N/A", "N/A"]
             if self.__status is PlayerStatus.IS_NOT_REGISTERED:
                 self.__status = PlayerStatus.IS_REGISTERED
                 self.__rank = 1
-            self._hasOwnAccount = False
+            self.__hasOwnAccount = False
             return updated
         updated = await self._addCharacters(charList)
         if updated:
             if self.__status is PlayerStatus.IS_NOT_REGISTERED:
                 self.__status = PlayerStatus.IS_REGISTERED
                 self.__rank = 1
-            self._hasOwnAccount = True
+            self.__hasOwnAccount = True
         return updated
 
     async def _addCharacters(self, charList):
@@ -284,7 +294,7 @@ class Player():
                         if p != self:
                             raise CharAlreadyExists(currName, p.id)
                     newIds[faction-1] = currId
-                    updated = updated or newIds[faction - 1] != self._igIds[faction-1]
+                    updated = updated or newIds[faction - 1] != self.__igIds[faction-1]
                     newNames[faction - 1] = jdata["character_list"][0]["name"]["first"]
             except IndexError:
                 # Should not happen, we checked earlier
@@ -298,10 +308,10 @@ class Player():
             if newIds[i] == 0:
                 raise CharMissingFaction(cfg.factions[i+1])
         if updated:
-            self._igIds = newIds.copy()
-            self._igNames = newNames.copy()
-            for i in range(len(self._igIds)):
-                _namesChecking[i][self._igIds[i]] = self
+            self.__igIds = newIds.copy()
+            self.__igNames = newNames.copy()
+            for i in range(len(self.__igIds)):
+                _namesChecking[i][self.__igIds[i]] = self
         return updated
 
 
@@ -311,7 +321,6 @@ class ActivePlayer:
 
     def __init__(self, player, team):
         self.__player = player
-        self.__player._active = self
         self.__illegalWeapons = dict()
         self.__kills = 0
         self.__deaths = 0
@@ -319,7 +328,23 @@ class ActivePlayer:
         self.__score = 0
         self.__team = team
         self.__account = None
-        self.__player.onPicked()
+        self.__player.onPicked(self)
+
+    @classmethod
+    def newFromData(cls, data, team):
+        try:
+            player = getPlayer(data["discord_id"])
+        except ElementNotFound:
+            player = Player(data["discord_id"], "unknown")
+        obj = cls(player, team)
+        obj.__score = data["score"]
+        obj.__net = data["net"]
+        obj.__deaths = data["deaths"]
+        obj.__kills = data["kills"]
+        obj.rank = data["rank"]
+        obj.__illWeaponsFromData(data["ill_weapons"])
+        return obj
+
 
     def clean(self):
         self.__player.onPlayerClean()
@@ -332,7 +357,7 @@ class ActivePlayer:
                 "net": self.__net,
                 "deaths": self.__deaths,
                 "kills": self.__kills,
-                "rank": self.__player.rank
+                "rank": self.rank
                 }
         return data
 
@@ -345,6 +370,17 @@ class ActivePlayer:
             data.append(doc)
         return data
 
+    def __illWeaponsFromData(self, data):
+        for weapDoc in data:
+            self.__illegalWeapons[weapDoc["weap_id"]] = weapDoc["kills"]
+
+    @property
+    def rank(self):
+        return self.__player.rank
+
+    @rank.setter
+    def rank(self, value):
+        self.__player.rank = value
 
     @property
     def name(self):
@@ -367,8 +403,7 @@ class ActivePlayer:
         fakePlayer = getPlayer(accountId)
         if fakePlayer is None:
             raise AccountNotFound(accountId)
-        self.__player._igNames = fakePlayer.igNames.copy()
-        self.__player._igIds = fakePlayer.igIds.copy()
+        self.__player.copyIgInfo(fakePlayer)
 
     def onMatchReady(self):
         self.__player.onMatchReady()
@@ -392,13 +427,13 @@ class ActivePlayer:
     def igId(self):
         faction = self.__team.faction
         if faction != 0:
-            return self.__player._igIds[faction-1]
+            return self.__player.igIds[faction-1]
 
     @property
     def igName(self):
         faction = self.__team.faction
         if faction != 0:
-            return self.__player._igNames[faction-1]
+            return self.__player.igNames[faction-1]
 
     @property
     def account(self):